requirements.txt
streamlit==1.32.2
pandas==2.2.1
numpy==1.26.4
plotly==5.20.0
pillow==10.2.0
